Add an optional field delimiter character to the "bundle" program.
Also, totally re-work its logic to avoid extra separator characters between
or following a field (the previous logic was lazy and always appended a
separator to the output line).  Getting it "right" is a lot more complicated
than doing it the old way.

One other benefit is that the new logic will also remove runs of blank lines.